+++ packages/collector/test/tracing/protocols/graphql/apolloServer.js
@@ -164,6 +164,7 @@ function pinoLogAndResolve(logMsg, value) {
   pinoLogger.warn(logMsg);
   return Promise.resolve(value);
 }
+
 function pinoLogAndReject(logMsg, error) {
   pinoLogger.warn(logMsg);
   return Promise.reject(error);